LLM thin client/requirements.txt
pymupdf
openai
google-generativeai
anthropic
cohere
requests
